+++ docs/docusaurus.config.js
@@ -224,17 +224,20 @@ const config = {
           },
         },
       },
+      algolia: {
+        appId: 'VVETP7QCVE',
+        apiKey: '167213b8ce51cc7ff9a804df130657e5',
+        indexName: 'ignite-cli',
+        contextualSearch: true,
+
+        // ↓ - To remove if `contextualSearch` versioning search works (to use if not)
+        // exclusionPatterns: [
+        //     'https://docs.ignite.com/v0.25.2/**',
+        //     'https://docs.ignite.com/nightly/**',
+        // ]
+      },
     }),
   plugins: [
-    [
-      require.resolve("@cmfcmf/docusaurus-search-local"),
-      {
-        indexDocs: true,
-        indexBlog: false,
-        indexPages: true,
-        language: "en",
-      },
-    ],
     [
       "@docusaurus/plugin-client-redirects",
       {