+++ ignite/pkg/cosmosanalysis/app/testdata/modules/external_variable/app.go
@@ -1,18 +1,31 @@
 package app
 
 import (
-	"app/modules"
-
-	"cosmossdk.io/client/v2/autocli"
-	"github.com/cosmos/cosmos-sdk/client"
+	"github.com/cosmos/cosmos-sdk/server/api"
+	"github.com/cosmos/cosmos-sdk/server/config"
+	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/types/module"
+	abci "github.com/tendermint/tendermint/abci/types"
+
+	"app/modules"
 )
 
 var ModuleBasics = module.NewBasicManager(modules.Basic...)
 
 type Foo struct{}
 
-func (Foo) Name() string                                    { return "foo" }
-func (Foo) InterfaceRegistry() codectypes.InterfaceRegistry { return nil }
-func (Foo) TxConfig() client.TxConfig                       { return nil }
-func (Foo) AutoCliOpts() autocli.AppOptions                 { return autocli.AppOptions{} }
+func (Foo) Name() string {
+	return "foo"
+}
+
+func (Foo) BeginBlocker(sdk.Context, abci.RequestBeginBlock) abci.ResponseBeginBlock {
+	return abci.ResponseBeginBlock{}
+}
+
+func (Foo) EndBlocker(sdk.Context, abci.RequestEndBlock) abci.ResponseEndBlock {
+	return abci.ResponseEndBlock{}
+}
+
+func (Foo) RegisterAPIRoutes(apiSvr *api.Server, apiConfig config.APIConfig) {
+	_ = apiSvr.ClientCtx
+}