+++ changelog.md
@@ -9,9 +9,11 @@
 ### Changes
 
 - [#3444](https://github.com/ignite/cli/pull/3444) Add support for ICS chains in ts-client generation
+- [#3494](https://github.com/ignite/cli/pull/3494) bump `cosmos-sdk` and `cometbft` versions
 
 ### Fixes
 
+- [#3497](https://github.com/ignite/cli/pull/3497) Use corret bank balance query url in faucet openapi
 - [#3481](https://github.com/ignite/cli/pull/3481) Use correct checksum format in release checksum file
 - [#3470](https://github.com/ignite/cli/pull/3470) Prevent overriding minimum-gas-prices with default value
 
@@ -76,6 +78,7 @@
 - [#3244](https://github.com/ignite/cli/pull/3244) Update `actions.yml` for resolving deprecation message
 - [#3337](https://github.com/ignite/cli/pull/3337) Remove `pkg/openapiconsole` import from scaffold template.
 - [#3337](https://github.com/ignite/cli/pull/3337) Register`nodeservice` gRPC in `app.go` template.
+- [#3455](https://github.com/ignite/cli/pull/3455) Bump `cosmos-sdk` to `v0.47.1`
 - [#3434](https://github.com/ignite/cli/pull/3434) Detect app wiring implementation.
 - [#3445](https://github.com/ignite/cli/pull/3445) refactor: replace `github.com/ghodss/yaml` with `sigs.k8s.io/yaml`